Refactor StreamWriter and extract encoding process (#3100)
Summary:
Pull Request resolved: https://github.com/pytorch/audio/pull/3100

Refactor StreamWriter and move OutputStream to dedicated source, then
split them into separate audio/video class.

Reviewed By: nateanl

Differential Revision: D43587337

fbshipit-source-id: 071b0bea57511ac5dc35c8adb76f2c611586f199
